Feature 2: add readme
