+++ README.md
@@ -1,1 +1,7 @@
 # ova-plan-api
+
+## How to run
+
+To run demo just execute `make run`.
+
+For other options see Makefile
